+++ NEWS.md
@@ -3,6 +3,7 @@
 ## New features
 
 * add `explore_col()` for simple bar plots without aggregation
+* add `yyyymm_calc()` for calculation with periods (format yyyymm)
 
 # explore 1.3.2